refactor(auth): Merge test helpers into module scope
Drop the separate pub(crate) mod test_support submodule.
TEST_PASSWORD and mock_config are now #[cfg(test)] items
directly in mod.rs; child modules access them via the
private ancestor path (super::super::mock_config) without
needing pub(crate).

Also set up Python Keystone in CI so the cross-verify
tests run for real instead of being silently skipped:
clone openstack/keystone, install bcrypt + cryptography,
and set KEYSTONE_PYTHON_CHECKOUT for the unit test step.

Note: This commit was done with the help of AI
