+++ .github/workflows/ci.yml
@@ -57,7 +57,19 @@ jobs:
       - name: Rust Cache
         uses: swatinem/rust-cache@98c8021b550208e191a6a3145459bfc9fb29c4c0 # v2.8.0
 
+      - name: Set up Python for cross-verification tests
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.x'
+
+      - name: Clone Python Keystone and install password hashing deps
+        run: |
+          git clone --depth=1 https://github.com/openstack/keystone.git /tmp/keystone-python
+          pip install bcrypt cryptography
+
       - name: Run tests (unit)
+        env:
+          KEYSTONE_PYTHON_CHECKOUT: /tmp/keystone-python
         run: cargo nextest run
 
       - name: Run tests (sqlite)
