+++ crates/core/src/common/password_hashing/bcrypt.rs
@@ -53,7 +53,7 @@ impl PasswordHasher for BcryptHasher {
 
 #[cfg(test)]
 mod tests {
-    use super::super::test_support::mock_config;
+    use super::super::mock_config;
     use super::super::{hash_password, verify_password};
     use openstack_keystone_config::PasswordHashingAlgo;
     use rand::distr::{Alphanumeric, SampleString};
