+++ crates/core/src/common/password_hashing/pbkdf2.rs
@@ -158,7 +158,7 @@ fn pbkdf2_hmac_sha512(
 
 #[cfg(test)]
 mod tests {
-    use super::super::test_support::{TEST_PASSWORD, mock_config};
+    use super::super::{TEST_PASSWORD, mock_config};
     use super::super::{hash_password, verify_password};
     use openstack_keystone_config::PasswordHashingAlgo;
