+++ crates/core/src/common/password_hashing/scrypt.rs
@@ -110,7 +110,7 @@ impl PasswordHasher for ScryptHasher {
 
 #[cfg(test)]
 mod tests {
-    use super::super::test_support::{TEST_PASSWORD, mock_config};
+    use super::super::{TEST_PASSWORD, mock_config};
     use super::super::{hash_password, verify_password};
     use openstack_keystone_config::PasswordHashingAlgo;
